fixed and separated module completion
